reimplementation of the extraction of property fetching
